Refactor `tts_config`
