+++ src/main.rs
@@ -167,21 +167,6 @@ fn init_subscriber(level: tracing::Level) -> Result<(), SetGlobalDefaultError> {
     tracing::subscriber::set_global_default(subscriber)
 }
 
-fn tts_config(config: &Config, provider: &Provider) -> TTSConfig {
-    let mut other = HashMap::new();
-    if provider != &Provider::Google {
-        other.insert("seed".to_string(), json!(42));
-    }
-    TTSConfig {
-        voice: Some(config.voice.clone()),
-        output_format: config.audio_format.clone(),
-        speed: config.speed,
-        seed: config.seed,
-        other: Some(other),
-        language_code: config.language_code.clone(),
-    }
-}
-
 pub(crate) fn audio_format(config: &Config) -> String {
     config.audio_format.clone().unwrap_or("mp3".to_string())
 }
@@ -200,7 +185,6 @@ pub(crate) async fn build(
         .as_ref()
         .map(|p| Provider::from_str(p).unwrap());
     let provider = provider.unwrap_or(Provider::DeepInfra);
-    let tts_config = tts_config(config, &provider);
 
     let slides = slide::slides(input.to_str().unwrap());
     if slides.is_empty() {
@@ -214,8 +198,7 @@ pub(crate) async fn build(
         out_dir,
         &slides,
         cache,
-        &tts_config,
-        &config.model,
+        &config,
         &audio_ext,
     )
     .await;
